% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_articles.R
\name{extract_nyt_data}
\alias{extract_nyt_data}
\title{Extract responses from a single page}
\usage{
extract_nyt_data(term, begin_date, end_date, key)
}
\arguments{
\item{term}{search term}

\item{begin_date}{beginning date}

\item{end_date}{end date}

\item{key}{New York Times API key}
}
\value{
out the dataframe output of the NYT API response
}
\description{
Extract responses from a single page
}
