% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_articles.R
\name{get_request}
\alias{get_request}
\title{Form a request URL}
\usage{
get_request(term, begin_date, end_date, key, page = 1)
}
\arguments{
\item{term}{search term}

\item{begin_date}{beginning date}

\item{end_date}{end date}

\item{key}{New York Times API key}

\item{page}{page number. The default value is 1.}
}
\value{
a request URL
}
\description{
Form a request URL
}
