% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_articles.R
\name{slowly_extract}
\alias{slowly_extract}
\title{Make the extracting function work slowly}
\usage{
slowly_extract(term, begin_date, end_date, key, page = 1)
}
\arguments{
\item{term}{search term}

\item{begin_date}{beginning date}

\item{end_date}{end date}

\item{key}{New York Times API key}

\item{page}{page number. The default value is 1.}

\item{rate_limit}{The rate limit. The default limit is 6 seconds.}
}
\description{
Make the extracting function work slowly
}
