% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_articles.R
\name{slowly_extract_nyt_data}
\alias{slowly_extract_nyt_data}
\title{Make the extracting function work slowly}
\usage{
slowly_extract_nyt_data(rate_limit = 6)
}
\arguments{
\item{rate_limit}{The rate limit. The default limit is 6 seconds.}
}
\description{
Make the extracting function work slowly
}
