% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_articles.R
\name{get_content}
\alias{get_content}
\title{Get response content}
\usage{
get_content(term, begin_date, end_date, key)
}
\arguments{
\item{term}{search term}

\item{begin_date}{beginning date}

\item{end_date}{end date}

\item{key}{New York Times API key}
}
\value{
a dataframe output
}
\description{
Get response content
}
